+++ cmd/fts/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/textileio/powergate/cmd/powtasker/cmd"
+	"github.com/textileio/powergate/cmd/fts/cmd"
 )
 
 func main() {
